UI CHANGES

+++ src/main/java/com/gharharyali/org/web/rest/NurseryResource.java
@@ -1,33 +1,41 @@
 package com.gharharyali.org.web.rest;
 
-import com.gharharyali.org.service.NurseryService;
-import com.gharharyali.org.web.rest.errors.BadRequestAlertException;
-import com.gharharyali.org.service.dto.NurseryDTO;
-import com.gharharyali.org.service.dto.NurseryCriteria;
-import com.gharharyali.org.service.NurseryQueryService;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.List;
+import java.util.Optional;
 
-import io.github.jhipster.web.util.HeaderUtil;
-import io.github.jhipster.web.util.PaginationUtil;
-import io.github.jhipster.web.util.ResponseUtil;
+import javax.validation.Valid;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import javax.validation.Valid;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.StreamSupport;
+import com.gharharyali.org.security.SecurityUtils;
+import com.gharharyali.org.security.resource.NurserySecurityValidation;
+import com.gharharyali.org.service.NurseryQueryService;
+import com.gharharyali.org.service.NurseryService;
+import com.gharharyali.org.service.dto.NurseryCriteria;
+import com.gharharyali.org.service.dto.NurseryDTO;
+import com.gharharyali.org.web.rest.errors.BadRequestAlertException;
 
-import static org.elasticsearch.index.query.QueryBuilders.*;
+import io.github.jhipster.web.util.HeaderUtil;
+import io.github.jhipster.web.util.PaginationUtil;
+import io.github.jhipster.web.util.ResponseUtil;
 
 /**
  * REST controller for managing {@link com.gharharyali.org.domain.Nursery}.
@@ -65,6 +73,7 @@ public class NurseryResource {
         if (nurseryDTO.getId() != null) {
             throw new BadRequestAlertException("A new nursery cannot already have an ID", ENTITY_NAME, "idexists");
         }
+        NurserySecurityValidation.nurserySecurityValidation(nurseryDTO);
         NurseryDTO result = nurseryService.save(nurseryDTO);
         return ResponseEntity.created(new URI("/api/nurseries/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.getId().toString()))
@@ -86,6 +95,7 @@ public class NurseryResource {
         if (nurseryDTO.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        NurserySecurityValidation.nurserySecurityValidation(nurseryDTO);
         NurseryDTO result = nurseryService.save(nurseryDTO);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, nurseryDTO.getId().toString()))
@@ -102,6 +112,7 @@ public class NurseryResource {
     @GetMapping("/nurseries")
     public ResponseEntity<List<NurseryDTO>> getAllNurseries(NurseryCriteria criteria, Pageable pageable) {
         log.debug("REST request to get Nurseries by criteria: {}", criteria);
+        NurserySecurityValidation.nurserySecurityValidation(criteria);
         Page<NurseryDTO> page = nurseryQueryService.findByCriteria(criteria, pageable);
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(ServletUriComponentsBuilder.fromCurrentRequest(), page);
         return ResponseEntity.ok().headers(headers).body(page.getContent());
@@ -116,6 +127,7 @@ public class NurseryResource {
     @GetMapping("/nurseries/count")
     public ResponseEntity<Long> countNurseries(NurseryCriteria criteria) {
         log.debug("REST request to count Nurseries by criteria: {}", criteria);
+        NurserySecurityValidation.nurserySecurityValidation(criteria);
         return ResponseEntity.ok().body(nurseryQueryService.countByCriteria(criteria));
     }
 
@@ -128,6 +140,11 @@ public class NurseryResource {
     @GetMapping("/nurseries/{id}")
     public ResponseEntity<NurseryDTO> getNursery(@PathVariable Long id) {
         log.debug("REST request to get Nursery : {}", id);
+        NurseryCriteria criteria =new NurseryCriteria();
+        NurserySecurityValidation.nurserySecurityValidation(criteria);
+        if(nurseryQueryService.countByCriteria(criteria)<1) {
+        	throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
         Optional<NurseryDTO> nurseryDTO = nurseryService.findOne(id);
         return ResponseUtil.wrapOrNotFound(nurseryDTO);
     }
@@ -141,6 +158,11 @@ public class NurseryResource {
     @DeleteMapping("/nurseries/{id}")
     public ResponseEntity<Void> deleteNursery(@PathVariable Long id) {
         log.debug("REST request to delete Nursery : {}", id);
+        NurseryCriteria criteria =new NurseryCriteria();
+        NurserySecurityValidation.nurserySecurityValidation(criteria);
+        if(nurseryQueryService.countByCriteria(criteria)<1) {
+        	throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
         nurseryService.delete(id);
         return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString())).build();
     }
@@ -156,6 +178,7 @@ public class NurseryResource {
     @GetMapping("/_search/nurseries")
     public ResponseEntity<List<NurseryDTO>> searchNurseries(@RequestParam String query, Pageable pageable) {
         log.debug("REST request to search for a page of Nurseries for query {}", query);
+        query=query.concat("username="+(SecurityUtils.getCurrentUserLogin().get()));
         Page<NurseryDTO> page = nurseryService.search(query, pageable);
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(ServletUriComponentsBuilder.fromCurrentRequest(), page);
         return ResponseEntity.ok().headers(headers).body(page.getContent());
